+++ src/widgets/distribution/settings.ts
@@ -22,7 +22,7 @@ export const settings: IWidgetSettings = makeSettings({
         initDataSets: [{viewType: 'DISTRIBUTION'}],
         canAdd: false,
         settings: [
-            makeString('name', 'Название'),
+            ...settingsPresets.dataSourceName,
             makeColor('color', ' Цвет'),
             makeList<ChartType>('chartType', 'Вид', 'LINE', ChartTypeValues),
             makeSettingsGroup('lineStyle', 'Стиль линии', [