+++ src/widgets/spline/settings.ts
@@ -40,7 +40,7 @@ export const settings: IWidgetSettings = makeSettings({
         initDataSets: [{viewType: 'DYNAMIC'}],
         canAdd: true,
         settings: [
-            makeString('name', 'Название'),
+            ...settingsPresets.dataSourceName,
             makeColor('color', ' Цвет'),
             makeList<ChartType>('chartType', 'Вид', 'LINE', ChartTypeValues),
             makeSettingsGroup('lineStyle', 'Стиль линии', [